requirements.txt
flask
cryptography
PyJWT
qrcode[pil]
